Move dataset folder to root of supergraph
