+++ archai/datasets/cv/augmentation/augmented_data_loader_provider.py
@@ -28,8 +28,8 @@ from archai.common.logger import Logger
 logger = Logger(source=__name__)
 from archai.common.config import Config
 from archai.datasets.cv.augmentation.augmentation_policy import add_named_augs
-from archai.supergraph.utils.datasets.dataset_provider import DatasetProvider, get_provider_type
-from archai.supergraph.utils.datasets.limit_dataset import DatasetLike, LimitDataset
+from archai.supergraph.datasets.dataset_provider import DatasetProvider, get_provider_type
+from archai.supergraph.datasets.limit_dataset import DatasetLike, LimitDataset
 from archai.datasets.distributed_stratified_sampler import DistributedStratifiedSampler
 from archai.supergraph.utils import apex_utils
