+++ archai/supergraph/algos/manual/manual_evaluater.py
@@ -19,7 +19,7 @@ from archai.supergraph.utils.trainer import Trainer
 from archai.common.config import Config
 from archai.common.logger import Logger
 logger = Logger(source=__name__)
-from archai.supergraph.utils.datasets import data
+from archai.supergraph.datasets import data
 from archai.supergraph.nas.model_desc import ModelDesc
 from archai.supergraph.nas.model_desc_builder import ModelDescBuilder
 from archai.supergraph.nas import nas_utils
