+++ archai/supergraph/datasets/providers/cifar100_provider.py
@@ -8,14 +8,13 @@ from torch.utils.data.dataset import Dataset
 
 import torchvision
 from torchvision.transforms import transforms
-from torch.utils.data import ConcatDataset
 
-from archai.supergraph.utils.datasets.dataset_provider import DatasetProvider, ImgSize, register_dataset_provider, TrainTestDatasets
+from archai.supergraph.datasets.dataset_provider import DatasetProvider, ImgSize, register_dataset_provider, TrainTestDatasets
 from archai.common.config import Config
 from archai.common import utils
 
 
-class SvhnProvider(DatasetProvider):
+class Cifar100Provider(DatasetProvider):
     def __init__(self, conf_dataset:Config):
         super().__init__(conf_dataset)
         self._dataroot = utils.full_path(conf_dataset['dataroot'])
@@ -26,21 +25,18 @@ class SvhnProvider(DatasetProvider):
         trainset, testset = None, None
 
         if load_train:
-            trainset = torchvision.datasets.SVHN(root=self._dataroot, split='train',
+            trainset = torchvision.datasets.CIFAR100(root=self._dataroot, train=True,
                 download=True, transform=transform_train)
-            extraset = torchvision.datasets.SVHN(root=self._dataroot, split='extra',
-                download=True, transform=transform_train)
-            trainset = ConcatDataset([trainset, extraset])
         if load_test:
-            testset = torchvision.datasets.SVHN(root=self._dataroot, split='test',
+            testset = torchvision.datasets.CIFAR100(root=self._dataroot, train=False,
                 download=True, transform=transform_test)
 
         return trainset, testset
 
     @overrides
     def get_transforms(self, img_size:ImgSize)->tuple:
-        MEAN = [0.4914, 0.4822, 0.4465]
-        STD = [0.2023, 0.1994, 0.20100]
+        MEAN = [0.507, 0.487, 0.441]
+        STD = [0.267, 0.256, 0.276]
         transf = [
             transforms.RandomCrop(32, padding=4),
             transforms.RandomHorizontalFlip()
@@ -56,4 +52,4 @@ class SvhnProvider(DatasetProvider):
 
         return train_transform, test_transform
 
-register_dataset_provider('svhn', SvhnProvider)
+register_dataset_provider('cifar100', Cifar100Provider)
