+++ archai/supergraph/datasets/providers/cifar10_provider.py
@@ -9,12 +9,11 @@ from torch.utils.data.dataset import Dataset
 import torchvision
 from torchvision.transforms import transforms
 
-from archai.supergraph.utils.datasets.dataset_provider import DatasetProvider, ImgSize, register_dataset_provider, TrainTestDatasets
+from archai.supergraph.datasets.dataset_provider import DatasetProvider, ImgSize, register_dataset_provider, TrainTestDatasets
 from archai.common.config import Config
 from archai.common import utils
 
-
-class Cifar100Provider(DatasetProvider):
+class Cifar10Provider(DatasetProvider):
     def __init__(self, conf_dataset:Config):
         super().__init__(conf_dataset)
         self._dataroot = utils.full_path(conf_dataset['dataroot'])
@@ -25,18 +24,18 @@ class Cifar100Provider(DatasetProvider):
         trainset, testset = None, None
 
         if load_train:
-            trainset = torchvision.datasets.CIFAR100(root=self._dataroot, train=True,
-                download=True, transform=transform_train)
+            trainset = torchvision.datasets.CIFAR10(root=self._dataroot,
+                train=True, download=True, transform=transform_train)
         if load_test:
-            testset = torchvision.datasets.CIFAR100(root=self._dataroot, train=False,
-                download=True, transform=transform_test)
+            testset = torchvision.datasets.CIFAR10(root=self._dataroot,
+                train=False, download=True, transform=transform_test)
 
         return trainset, testset
 
     @overrides
     def get_transforms(self, img_size:ImgSize)->tuple:
-        MEAN = [0.507, 0.487, 0.441]
-        STD = [0.267, 0.256, 0.276]
+        MEAN = [0.49139968, 0.48215827, 0.44653124]
+        STD = [0.24703233, 0.24348505, 0.26158768]
         transf = [
             transforms.RandomCrop(32, padding=4),
             transforms.RandomHorizontalFlip()
@@ -52,4 +51,4 @@ class Cifar100Provider(DatasetProvider):
 
         return train_transform, test_transform
 
-register_dataset_provider('cifar100', Cifar100Provider)
+register_dataset_provider('cifar10', Cifar10Provider)
