+++ archai/supergraph/datasets/providers/imagenet_folder.py
