+++ archai/supergraph/datasets/providers/imagenet_provider.py
@@ -15,7 +15,7 @@ from PIL import Image
 
 from overrides import overrides, EnforceOverrides
 
-from archai.supergraph.utils.datasets.dataset_provider import DatasetProvider, ImgSize, register_dataset_provider, TrainTestDatasets
+from archai.supergraph.datasets.dataset_provider import DatasetProvider, ImgSize, register_dataset_provider, TrainTestDatasets
 from archai.common.config import Config
 from archai.common import utils
 from archai.datasets.cv.transforms.lighting import Lighting
