+++ archai/supergraph/datasets/providers/mit67_provider.py
@@ -10,12 +10,12 @@ from torch.utils.data.dataset import Dataset
 import torchvision
 from torchvision.transforms import transforms
 
-from archai.supergraph.utils.datasets.dataset_provider import DatasetProvider, ImgSize, register_dataset_provider, TrainTestDatasets
+from archai.supergraph.datasets.dataset_provider import DatasetProvider, ImgSize, register_dataset_provider, TrainTestDatasets
 from archai.common.config import Config
 from archai.common import utils
 
 
-class StanfordCarsProvider(DatasetProvider):
+class Mit67Provider(DatasetProvider):
     def __init__(self, conf_dataset:Config):
         super().__init__(conf_dataset)
         self._dataroot = utils.full_path(conf_dataset['dataroot'])
@@ -26,10 +26,10 @@ class StanfordCarsProvider(DatasetProvider):
         trainset, testset = None, None
 
         if load_train:
-            trainpath = os.path.join(self._dataroot, 'stanfordcars', 'train')
+            trainpath = os.path.join(self._dataroot, 'mit67', 'train')
             trainset = torchvision.datasets.ImageFolder(trainpath, transform=transform_train)
         if load_test:
-            testpath = os.path.join(self._dataroot, 'stanfordcars', 'test')
+            testpath = os.path.join(self._dataroot, 'mit67', 'test')
             testset = torchvision.datasets.ImageFolder(testpath, transform=transform_test)
 
         return trainset, testset
@@ -41,7 +41,7 @@ class StanfordCarsProvider(DatasetProvider):
         if isinstance(img_size, int):
             img_size = (img_size, img_size)
 
-        # TODO: update MEAN, STD, currently mit67 values
+        # MEAN, STD computed for mit67
         MEAN = [0.4893, 0.4270, 0.3625]
         STD = [0.2631, 0.2565, 0.2582]
 
@@ -59,7 +59,6 @@ class StanfordCarsProvider(DatasetProvider):
 
         margin_size = (int(img_size[0] + img_size[0]*0.1), int(img_size[1] + img_size[1]*0.1))
         test_transf = [transforms.Resize(margin_size), transforms.CenterCrop(img_size)]
-        #test_transf = [transforms.Resize(img_size)]
 
         normalize = [
             transforms.ToTensor(),
@@ -71,4 +70,4 @@ class StanfordCarsProvider(DatasetProvider):
 
         return train_transform, test_transform
 
-register_dataset_provider('stanfordcars', StanfordCarsProvider)
+register_dataset_provider('mit67', Mit67Provider)
