+++ archai/supergraph/datasets/providers/person_coco_provider.py
@@ -10,12 +10,12 @@ from torch.utils.data.dataset import Dataset
 import torchvision
 from torchvision.transforms import transforms
 
-from archai.supergraph.utils.datasets.dataset_provider import DatasetProvider, ImgSize, register_dataset_provider, TrainTestDatasets
+from archai.supergraph.datasets.dataset_provider import DatasetProvider, ImgSize, register_dataset_provider, TrainTestDatasets
 from archai.common.config import Config
 from archai.common import utils
 
 
-class Food101Provider(DatasetProvider):
+class PersonCocoProvider(DatasetProvider):
     def __init__(self, conf_dataset:Config):
         super().__init__(conf_dataset)
         self._dataroot = utils.full_path(conf_dataset['dataroot'])
@@ -26,24 +26,25 @@ class Food101Provider(DatasetProvider):
         trainset, testset = None, None
 
         if load_train:
-            trainpath = os.path.join(self._dataroot, 'food-101', 'train')
+            trainpath = os.path.join(self._dataroot, 'person_coco', 'train')
             trainset = torchvision.datasets.ImageFolder(trainpath, transform=transform_train)
         if load_test:
-            testpath = os.path.join(self._dataroot, 'food-101', 'test')
+            testpath = os.path.join(self._dataroot, 'person_coco', 'test')
             testset = torchvision.datasets.ImageFolder(testpath, transform=transform_test)
 
         return trainset, testset
 
     @overrides
     def get_transforms(self, img_size:ImgSize)->tuple:
+        # TODO: update MEAN, STD, currently mit67 values
+        MEAN = [0.4893, 0.4270, 0.3625]
+        STD = [0.2631, 0.2565, 0.2582]
 
-        print(f'IMG SIZE: {img_size}')
+        # transformations match that in
+        # https://github.com/antoyang/NAS-Benchmark/blob/master/DARTS/preproc.py
         if isinstance(img_size, int):
             img_size = (img_size, img_size)
 
-        # TODO: Need to rethink the food101 transforms
-        MEAN = [0.5451, 0.4435, 0.3436]
-        STD = [0.2171, 0.2251, 0.2260] # TODO: should be [0.2517, 0.2521, 0.2573]
         train_transf = [
             transforms.RandomResizedCrop(img_size, scale=(0.75, 1)),
             transforms.RandomHorizontalFlip(),
@@ -54,9 +55,9 @@ class Food101Provider(DatasetProvider):
                 hue=0.2)
         ]
 
-        # food101 has images of varying sizes and are ~512 each side
         margin_size = (int(img_size[0] + img_size[0]*0.1), int(img_size[1] + img_size[1]*0.1))
         test_transf = [transforms.Resize(margin_size), transforms.CenterCrop(img_size)]
+        #test_transf = [transforms.Resize(72), transforms.CenterCrop(img_size)]
 
         normalize = [
             transforms.ToTensor(),
@@ -68,4 +69,4 @@ class Food101Provider(DatasetProvider):
 
         return train_transform, test_transform
 
-register_dataset_provider('food101', Food101Provider)
+register_dataset_provider('person_coco', PersonCocoProvider)
