+++ archai/supergraph/nas/arch_trainer.py
@@ -13,7 +13,7 @@ from archai.common.config import Config
 from archai.common import utils
 from archai.supergraph.nas.model import Model
 from archai.supergraph.utils.trainer import Trainer
-from archai.supergraph.utils.datasets import data
+from archai.supergraph.datasets import data
 from archai.supergraph.nas.vis_model_desc import draw_model_desc
 from archai.supergraph.utils.checkpoint import CheckPoint
