+++ archai/supergraph/nas/searcher.py
@@ -14,7 +14,7 @@ from archai.supergraph.nas.model_desc_builder import ModelDescBuilder
 from archai.supergraph.nas.arch_trainer import TArchTrainer
 from archai.supergraph.utils.trainer import Trainer
 from archai.supergraph.nas.model_desc import ModelDesc
-from archai.supergraph.utils.datasets import data
+from archai.supergraph.datasets import data
 from archai.supergraph.nas.model import Model
 from archai.supergraph.utils.metrics import Metrics
 from archai.supergraph.nas.finalizers import Finalizers
