+++ archai/supergraph/utils/trainer.py
@@ -17,7 +17,7 @@ from archai.common.config import Config
 from archai.common import utils
 from archai.supergraph.utils import ml_utils
 from archai.common.common import logger
-from archai.supergraph.utils.datasets import data
+from archai.supergraph.datasets import data
 from archai.supergraph.utils.checkpoint import CheckPoint
 from archai.common.apex_utils import ApexUtils
 from archai.supergraph.utils.multi_optim import MultiOptim, OptimSched
